+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Changelog starting from v0.5 (2018/09/18)
 - Sell page: we can pay with **two payment methods**.
 - the Commands page shows **all cards to command**, in addition to them grouped by supplier. This makes it easier to download a CSV and send it to a supplier.
 - added a button to collapse the left menu. We can now work in "full screen". Especially useful for the Raspberry Pis and views with a lot of information (Reception, lists…).
+- We can choose what books to show on the ABStock catalogue from within Abelujo (see new page `/catalogue-selection`).
 
 ### Other enhancements